added resetUserSelection() to avoid negative stock levels
a fringe case was allowing the full stock to be added to the basket without reseting the count, allowing users to leave stock levels below 0. This was noticed while stress-testing the page and corrected by adding a function dedicated to resetting the user selection in order to account for fringe cases.